undoing src and main
