+++ src/aq.py
@@ -7,9 +7,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 from tqdm.auto import trange
 
-from src.inference import FinalizedQuantizedLinear
 from src.kmeans import find_nearest_cluster, fit_faiss_kmeans, fit_kmeans, fit_kmeans_1d
-from src.utils import _dequantize_weight, ellipsis, maybe_script, pack_int_data
+from src.utils import ellipsis, maybe_script
 
 
 class QuantizedLinear(nn.Module):
@@ -22,32 +21,6 @@ class QuantizedLinear(nn.Module):
     def forward(self, input: torch.Tensor):
         # TODO[aqlm] this can be optimized! maybe integrate with QuantizedLinear?
         return F.linear(input, self.quantized_weight(), self.bias)
-
-    @torch.no_grad()
-    def finalize(self) -> FinalizedQuantizedLinear:
-        assert self.quantized_weight is not None, "Can't finalize an unprocessed layer"
-        finalized_quantized_linear = FinalizedQuantizedLinear(
-            self.in_features,
-            self.out_features,
-            self.quantized_weight.in_group_size,
-            self.quantized_weight.out_group_size,
-            self.quantized_weight.num_codebooks,
-            self.quantized_weight.nbits_per_codebook,
-            self.bias is not None,
-            device=self.quantized_weight.codes.device,
-        )
-
-        state_dict = {
-            "codes": pack_int_data(self.quantized_weight.codes.clone(), self.quantized_weight.nbits_per_codebook),
-            "codebooks": self.quantized_weight.get_codebooks().clone(),
-            "scales": self.quantized_weight.get_scales().clone(),
-        }
-        if self.bias is not None:
-            state_dict["bias"] = self.bias.clone()
-
-        finalized_quantized_linear.load_state_dict(state_dict)
-
-        return finalized_quantized_linear
 
 
 class QuantizedWeight(nn.Module):
@@ -387,6 +360,36 @@ def beam_search_optimal_codes(
 
                 progressbar.desc = info
     return beam_codes[0]
+
+
+@maybe_script
+def _dequantize_weight(
+    codes: torch.Tensor, codebooks: torch.Tensor, scales: Optional[torch.Tensor] = None
+) -> torch.Tensor:
+    """
+    Decode float weights from quantization codes. Differentiable.
+    :param codes: tensor of integer quantization codes, shape [*dims, num_out_groups, num_in_groups, num_codebooks]
+    :param codebooks: tensor of vectors for each quantization code, [num_codebooks, codebook_size, out_group_size, in_group_size]
+    :param scales: weight will be multiplied by this factor, must be broadcastble with [*dims, out_groups, num_in_groups, out_group_size, in_group_size]
+    :return: reconstructed weight tensor of shape [*dims, num_in_groups*group_size]
+    """
+    num_out_groups, num_in_groups, num_codebooks = codes.shape[-3:]
+    num_codebooks, codebook_size, out_group_size, in_group_size = codebooks.shape
+    out_features = num_out_groups * out_group_size
+    in_features = num_in_groups * in_group_size
+    codebook_offsets = torch.arange(
+        0, num_codebooks * codebook_size, codebook_size, device=codes.device
+    )  # shape: [num_codebooks]
+    reconstructed_weight_flat = F.embedding_bag(
+        codes.flatten(0, -2) + codebook_offsets, codebooks.flatten(0, 1).flatten(-2, -1), mode="sum"
+    )  # [prod(dims) * num_out_groups * num_in_groups, out_group_size * in_group_size]
+
+    reconstructed_weight_groupwise = reconstructed_weight_flat.view(
+        list(codes.shape[:-3]) + [num_out_groups, num_in_groups, out_group_size, in_group_size]
+    )
+    if scales is not None:
+        reconstructed_weight_groupwise = reconstructed_weight_groupwise.mul(scales)
+    return reconstructed_weight_groupwise.swapaxes(-3, -2).reshape(list(codes.shape[:-3]) + [out_features, in_features])
 
 
 @maybe_script
